Color increase 4 attempt:  What about reading the COLUBK into, say x, increase it and write back to find out why the previous direct "inc COLUBK" didn't work? Start2: 	ldx COLUBK 	inx 	stx COLUBK 	jmp Start2
Interestingly, the Stella disassembler lists "ldx COLUBK" as "ldx INPT1" - is it possible we can't read COLUBK at all but, for reading, the adress gives up something entirely different? And this is why the above direct "inc COLUBK" didn't work either as it ALWAYS got INPT1, which is always 9 (1001), and it just stored this increased-to-10 value in each loop back to the bg!